requirements.txt
numpy
PILLOW
tensorflow
streamlit
pickle